+++ src/app/editAccountInfo/[id]/page.tsx
@@ -3,8 +3,8 @@ import editProfileApi, { EditRes } from "@/api/editProfileApi";
 import DaumPostcode, { AddressData } from "@/components/common/daumPostcode";
 import ModalFilter from "@/components/modal/modalFilter";
 import { BankOptions } from "@/components/common/selectOption";
-import useDecodedToken from "@/hooks/useDecodedToken";
-import useInput from "@/hooks/useInput";
+import useDecodedToken from "@/hooks/common/useDecodedToken";
+import useInput from "@/hooks/common/useInput";
 import useAuthStore from "@/store/useAuthStore";
 import { AccountInfoMain } from "@/styles/joinStyle";
 import { useRouter } from "next/navigation";
