+++ src/app/find/password/page.tsx
@@ -3,8 +3,8 @@ import pwFindApi from "@/api/passwordFindApi";
 import resetPasswordApi from "@/api/resetPasswordApi";
 import verifyCodeApi from "@/api/verifyCodeApi";
 import FindForm from "@/components/form/findForm";
-import useInput from "@/hooks/useInput";
-import useValid from "@/hooks/useValid";
+import useInput from "@/hooks/common/useInput";
+import useValid from "@/hooks/common/useValid";
 import { ErrorMessage, FindMain } from "@/styles/joinStyle";
 import { useRouter } from "next/navigation";
 import { FormEvent, useState } from "react";
