+++ src/app/update/product/[id]/page.tsx
@@ -7,7 +7,7 @@ import {
   ProductStatus,
 } from "@/components/common/selectOption";
 import TagInput from "@/components/common/tagInput";
-import useInput from "@/hooks/useInput";
+import useInput from "@/hooks/common/useInput";
 import {
   BasicImg,
   ImgFile,
