+++ src/app/upload/groupPurchase/page.tsx
@@ -3,7 +3,7 @@ import groupUploadApi, { GroupReq } from "@/api/groupUploadApi";
 import uploadApi from "@/api/uploadApi";
 import { BankOptions, DeliveryOptions } from "@/components/common/selectOption";
 import TagInput from "@/components/common/tagInput";
-import useInput from "@/hooks/useInput";
+import useInput from "@/hooks/common/useInput";
 import {
   GroupForm,
   UploadMain,
