+++ src/components/common/header.tsx
@@ -15,8 +15,8 @@ import IconChat from "@/../public/assets/svgs/messages.svg";
 import { useRouter } from "next/navigation";
 import useAuthStore from "@/store/useAuthStore";
 import ModalFilter from "../modal/modalFilter";
-import useInput from "@/hooks/useInput";
-import useDecodedToken from "@/hooks/useDecodedToken";
+import useInput from "@/hooks/common/useInput";
+import useDecodedToken from "@/hooks/common/useDecodedToken";
 import useChatStore from "@/store/useChatStore";
 import chatExitApi from "@/api/chat/chatExitApi";
